requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
Flask==3.0.0
edge-tts==6.1.9